+++ lahta/.env
@@ -1,0 +1,5 @@
+BUCKET=lahta.contextmachine.online
+STORAGE=https://storage.yandexcloud.net/
+REGION=ru-central1
+AWS_ACCESS_KEY_ID=YCAJEWyfhMEdpJv5yhuXECVLT
+AWS_SECRET_ACCESS_KEY=YCPsGaw0zZXczhE-vtsG2_XYgM7yW1F_FqCg7de1
